Uhh updated TestNail, reworked a lot of things

+++ ConcreteTableInheritance/src/domain/Fastener.java
@@ -3,7 +3,7 @@ package domain;
 import java.util.ArrayList;
 
 public abstract class Fastener extends InventoryItem {
-	protected long length;
+	protected double length;
 	
 	public Fastener(ArrayList<Object> objectArray) {
 		super(objectArray);
@@ -11,7 +11,7 @@ public abstract class Fastener extends InventoryItem {
 
 
 
-	public long getLength() {
+	public double getLength() {
 		return this.length;
 	}
 }
